% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{snowflake_azure_auth}
\alias{snowflake_azure_auth}
\title{Authenticate to Snowflake via Azure Active Directory}
\usage{
snowflake_azure_auth(account, app_uri)
}
\arguments{
\item{account}{The Snowflake account identifier, e.g. \code{myaccount} or
\code{myaccount.snowflakecomputing.com}.}

\item{app_uri}{The Azure AD application URI for Snowflake, usually in the
form \verb{api://...}.}
}
\description{
When using Posit Workbench and Azure AD for single-sign on, this
function can be used to retrieve delegated credentials for Snowflake that
can be used with \code{\link[=snowflake_conn]{snowflake_conn()}}.
}
\examples{
\dontrun{
snowflake_azure_auth(
  "wda27712.us-east-1", "api://a21cefc2-bb68-4a19-a8b2-fce415ce9da4"
)
}
}
